Complete first quick attempt at Single Price Grid Component challenge.
I captured the basic essence of the design, but there are quite a few details I need to refine! It appears I will have to do a lot of changes in the media query, so I'm trying to think of more clever ways to avoid putting so much code in them.